Add top 100 tokens so that offline tool can tokens we choose to support in the future
Summary:
BG-3629

This diff makes it so that the offline tool can sign arbitrary tokens we might add in the future. I have also listed the top 100 tokens or so. This is so that they can be recognized by the offline tool incase we decide to add them

Reviewers: bigly, tyler, arik

Reviewed By: tyler, arik

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D8384

(cherry picked from commit 83f1704)